Address feedback.
- Allow SemanticRewriteSuite to take custom classpath.
- Move IntegrationPropertyTest to testkit.
- Move testkit classes to scalafix.testkit package.
