+++ scalafix-testkit/src/main/scala/scalafix/testkit/Command.scala
