+++ scalafix-tests/src/test/scala/scalafix/SemanticTests.scala
@@ -1,9 +1,13 @@
 package scalafix
 
-class SemanticTests extends SemanticRewriteSuite {
+class SemanticTests
+    extends testkit.SemanticRewriteSuite(
+      // classpath to use for compiling .source files.
+      // In the scalafix repo, the system property is populated from build.sbt
+      sys.props("sbt.paths.scalafixNsc.test.classes")) {
   // directory containing .source files
   val testDir = "scalafix-tests/src/test/resources"
-  DiffTest.testsToRun(new java.io.File(testDir)).foreach { dt =>
+  testkit.DiffTest.testsToRun(new java.io.File(testDir)).foreach { dt =>
     if (dt.skip) {
       ignore(dt.fullName) {}
     } else {
